Move getSourceCode to its own file
